#33 issue: change name model classes and ID filed type BigInteger change to Long

+++ src/main/java/kz/bars/familybudget/repository/RoleRepo.java
@@ -5,11 +5,9 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.math.BigInteger;
-
 @Repository
 @Transactional
-public interface RoleRepo extends JpaRepository<Role, BigInteger> {
+public interface RoleRepo extends JpaRepository<Role, Long> {
 
     Role findByRole(String role);
